+++ library/src/lapack/roclapack_syevj_heevj_batched.cpp
@@ -43,20 +43,21 @@ rocblas_status rocsolver_syevj_heevj_batched_impl(rocblas_handle handle,
 
     // memory workspace sizes:
     // size of temporary workspace
-    size_t size_Acpy, size_cosines, size_sines, size_top, size_bottom, size_completed;
+    size_t size_Acpy, size_cosines, size_sines, size_top, size_bottom, size_completed, size_norms;
 
     rocsolver_syevj_heevj_getMemorySize<true, T, S>(evect, uplo, n, batch_count, &size_Acpy,
                                                     &size_cosines, &size_sines, &size_top,
-                                                    &size_bottom, &size_completed);
+                                                    &size_bottom, &size_completed, &size_norms);
 
     if(rocblas_is_device_memory_size_query(handle))
         return rocblas_set_optimal_device_memory_size(handle, size_Acpy, size_cosines, size_sines,
-                                                      size_top, size_bottom, size_completed);
+                                                      size_top, size_bottom, size_completed,
+                                                      size_norms);
 
     // memory workspace allocation
-    void *Acpy, *cosines, *sines, *top, *bottom, *completed;
+    void *Acpy, *cosines, *sines, *top, *bottom, *completed, *norms;
     rocblas_device_malloc mem(handle, size_Acpy, size_cosines, size_sines, size_top, size_bottom,
-                              size_completed);
+                              size_completed, size_norms);
 
     if(!mem)
         return rocblas_status_memory_error;
@@ -67,12 +68,13 @@ rocblas_status rocsolver_syevj_heevj_batched_impl(rocblas_handle handle,
     top = mem[3];
     bottom = mem[4];
     completed = mem[5];
+    norms = mem[6];
 
     // execution
     return rocsolver_syevj_heevj_template<true, false, T>(
         handle, esort, evect, uplo, n, A, shiftA, lda, strideA, abstol, residual, max_sweeps,
         n_sweeps, W, strideW, info, batch_count, (T*)Acpy, (S*)cosines, (T*)sines,
-        (rocblas_int*)top, (rocblas_int*)bottom, (rocblas_int*)completed);
+        (rocblas_int*)top, (rocblas_int*)bottom, (rocblas_int*)completed, (S*)norms);
 }
 
 /*